#994 extract "/@db" part of DBPlugin to a separate class DBBrowserPlugin
* it has a totally different responsibility, and
* it could disable itself in prod mode
