+++ framework/src/play.plugins
@@ -3,6 +3,7 @@
 100:play.data.parsing.TempFilePlugin
 200:play.data.validation.ValidationPlugin
 300:play.db.DBPlugin
+301:play.db.DBBrowserPlugin
 400:play.db.jpa.JPAPlugin
 450:play.db.Evolutions
 500:play.i18n.MessagesPlugin
